requirements.txt
six
Faker
pillow
markdown
python-dateutil
pdf2image
stripe

django
django-environ
django-filter
django-allauth
django-resized
python-decouple
django-ckeditor
django-crispy-forms
crispy-bootstrap5
django-tinymce

django-browser-reload
stripe
django-jazzmin